app/pyproject.toml
[tool.poetry]
name = "backend application"
version = "0.0.1"
description = "backend application"
authors = ["Hiroshi Omata <op.homata@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "*"
django-environ = "*"
djangorestframework = "*"
djangorestframework-simplejwt = "*"
django-import-export = "*"
django-jazzmin = "*"
djoser = "*"
python-decouple = "*"
pandas = "*"
django-cors-headers = "*"
djangorestframework-api-key = "*"
gunicorn =  "*"
psycopg2-binary = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
django-extensions = "*"
django-debug-toolbar = "*"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
